separate prowler v3 and v4 parsers
